refactor: extract ChatMediaDelegate from ChatViewModel (verbatim move)
